test(queue): split bullmq-queue.integration.test.ts by concern (ps-ro3q)
Split bullmq-queue.integration.test.ts (1061 LOC) into 6 focused files
plus shared fixtures. Test count preserved (128 -> 128). All files <500
LOC.

- bullmq-queue-contract.integration.test.ts (110 LOC)
- bullmq-queue-connection.integration.test.ts (69 LOC)
- bullmq-queue-enqueue-branches.integration.test.ts (271 LOC)
- bullmq-queue-listing-branches.integration.test.ts (234 LOC)
- bullmq-queue-corruption-branches.integration.test.ts (263 LOC)
- bullmq-queue-worker-branches.integration.test.ts (187 LOC)
- helpers/bullmq-test-fixtures.ts (126 LOC)

Disable file-level parallelism for queue-integration in vitest.config.ts:
all 6 files share the same Valkey container, and running them in parallel
produced cross-file state collisions (BullMQ jobs leaking across queue
instances). Sequential file execution preserves the single-process
semantics the original single-file test enjoyed implicitly.
